rm tf/flax tests in /generation
